+++ src/Microsoft.DotNet.Helix/Sdk/Readme.md
@@ -331,17 +331,31 @@ Given a local folder `$(TestFolder)` containing `runtests.cmd`, this will run `r
   </PropertyGroup>
 
   <!--
-    XUnit v3 Runner
-      Enabling this will create one work item for each xunit v3 test project specified.
-      XUnit v3 tests are self-hosting executables and do not need an external console runner.
-      This is enabled by specifying one or more XUnitV3Project items.
+    Microsoft.Testing.Platform (MTP) Runner
+      Enabling this will create one Helix work item per MTP-based test project. This
+      covers MSTest 4.x with the MTP runner, xUnit v3 with MTP (the default for v3),
+      NUnit with the MTP runner, TUnit, and any custom MTP-based framework.
+
+      Each test project must reference Microsoft.Testing.Extensions.TrxReport so that
+      results can be reported as a TRX file (which arcade's reporter consumes natively).
+      Projects built with MSTest.Sdk, or with Microsoft.DotNet.Arcade.Sdk's XUnitV3
+      targets, get this reference implicitly.
+
+      For MSTest projects you can use the discoverability alias <MSTestProject .../>;
+      it is folded into @(MTPProject) before any other target runs.
   -->
   <ItemGroup>
-    <XUnitV3Project Include="..\tests\bar.Tests.csproj"/>
+    <!-- xUnit v3 / NUnit-MTP / TUnit / custom MTP -->
+    <MTPProject Include="..\tests\bar.Tests.csproj"/>
+
+    <!-- MSTest (folded into MTPProject) -->
+    <MSTestProject Include="..\tests\baz.Tests.csproj"/>
   </ItemGroup>
   <PropertyGroup>
-    <!-- Whether to use the Microsoft Testing Platform runner. Defaults to true. -->
-    <UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
+    <!-- Optional: customize the TRX filename produced by Microsoft.Testing.Extensions.TrxReport. Defaults to testResults.trx. -->
+    <MTPTrxReportFilename>testResults.trx</MTPTrxReportFilename>
+    <!-- Optional: per-work-item timeout (TimeSpan format). Defaults to 5 minutes. -->
+    <MTPWorkItemTimeout>00:05:00</MTPWorkItemTimeout>
   </PropertyGroup>
 
   <ItemGroup>
